Simplify KDA recurrence component tests
